requirements.txt
fastapi==0.115.12
gtts==2.5.4
google-cloud-texttospeech==2.26.0
openai==1.76.2
pandas==2.2.3
pydub==0.25.1
python-dotenv==1.1.0
uvicorn==0.34.2